Move common files to `/global/` directory
