+++ global/BrowserStorageManager.js
