+++ global/browserActions.js
